winegstreamer: Don't set our source pad active at all.
There's not any reason to do this here; this should be handled with state
change transitions. Since we start in the ready state, we shouldn't be active
anyway, and the only effect this will have is push_data() trying to run and
failing since the decodebin sink pad is flushing.