ntdll: Workaround for requesting thread affinity mask that isn't a subset of the system's mask
This patch is related to the game 'Divinity Original Sin 2' which crashes on startup, if the system only
has 4 or less logical cpus.

By Microsoft specification, requesting a thread affinity mask requires it be be a subset of the process affiniy mask.
However the game requests a mask F...FE (-2) which makes wine to return STATUS_INVALID_PARAMETER, which in turn  makes
the game hang.

This patch will make wine to adjust the mask in this case, which makes the game start correctly.

Signed-off-by: Roger Zoellner <zoellner.roger@gmail.com>