[local] wineserver: Handle MMCSS priorities in wineserver.
When a thread has its MMCSS priority set, we override whatever priority
it was assigned previously and set the priority directly without going
through any calculations. It reverts back when the MMCSS priority will
be unset. This is a dirty implementation of handling AVRT priority
requests without creating the MMCSS management service.

Signed-off-by: Kai Krakow <kai@kaishome.de>