[wip] tooling: Add patch collection tooling.
A work-in-progress list of applied patches is kept in `docs/patches` in
YAML format. The current design is a manually crafted inverse dependency
list. In the future it should generate this tree automatically from
staging definition files. The leafs of this tree are the patche series
applied to this repository.

Running `tools/mkpatchlist.rb` will generate a list of files suitable to
be fed to `git am`.

You need to point the tool to the staging patches repository by creating
a symlink to its patches directory, e.g.:

```bash
echo "/patches" >>.git/info/exclude
ln -s ../wine-staging/patches
```

Now run `tools/mkpatchlist.rb` to compile a list of patch files for
`git-am`. In the future this tool will be part of a toolchain to
recreate this wine branch from scratch on top of a vanilla wine branch.
Also in the future, it will list the Proton patches as individual patch
series to make crafting an individual wine version more easy. I'll then
push a wine patches repository forked from the staging patches
repository.

Signed-off-by: Kai Krakow <kai@kaishome.de>