kernel32: Launch 32-bit IL-only as 64-bit in WIN64
When using CreateProcessA to execute a 32-bit .NET/Mono application
(which is IL-only) the process is always launched as a 32-bit process
under Wine; but under Windows it is launched as a 64-bit process.

This patch changes Wine's behaviour to match Windows.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=46319
Signed-off-by: Brendan McGrath <brendan@redmandi.com>